requirements.txt
anyio==4.6.2
certifi==2024.8.30
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ollama==0.3.3
smmap==5.0.1
sniffio==1.3.1
